+++ ext/SerdeCSVExt/De.jl
@@ -1,19 +1,18 @@
 module DeCsv
 
-export deser_csv
-
 using ..ParCsv
 import Serde.to_deser
+import Serde
 
-function deser_csv(::Type{T}, x; kw...) where {T}
-    return to_deser(Vector{T}, parse_csv(x; kw...))
+function Serde.CSV.deser_csv(::Type{T}, x; kw...) where {T}
+    return to_deser(Vector{T}, Serde.CSV.parse_csv(x; kw...))
 end
 
-deser_csv(::Type{Nothing}, _) = nothing
-deser_csv(::Type{Missing}, _) = missing
+Serde.CSV.deser_csv(::Type{Nothing}, _) = nothing
+Serde.CSV.deser_csv(::Type{Missing}, _) = missing
 
-function deser_csv(f::Function, x; kw...)
-    object = parse_csv(x; kw...)
+function Serde.CSV.deser_csv(f::Function, x; kw...)
+    object = Serde.CSV.parse_csv(x; kw...)
     return to_deser(f(object), object)
 end
